// dhHelpdeskNG/DH.Helpdesk.WebMobile/src/app/models/caseTemplate/case-template-full.model.ts
export class CaseTemplateFullModel {
  id: number;
  caseType_Id: number;
  category_Id: number;
  department_Id: number;
  finishingCause_Id: number;
  noMailToNotifier: boolean;
  performerUser_Id: number;
  priority_Id: number;
  productArea_Id: number;
  project_Id: number;
  workingGroup_Id: number;
  caption: string;
  description: string;
  miscellaneous: string;
  name: string;
  reportedBy: string;
  text_External: string;
  text_Internal: string;
  orderNum: string;
  personsName: string;
  personsPhone: string;
  personsCellPhone: string;
  personsEmail: string;
  region_Id: number;
  oU_Id: number;
  place: string;
  userCode: string;
  system_Id: number;
  urgency_Id: number;
  impact_Id: number;
  invoiceNumber: string;
  referenceNumber: string;
  status_Id: number;
  stateSecondary_Id: number;
  verified: boolean;
  verifiedDescription: string;
  solutionRate: string;
  inventoryNumber: string;
  inventoryType: string;
  inventoryLocation: string;
  supplier_Id: number;
  agreedDate: string;
  available: string;
  cost: number;
  otherCost: number;
  currency: string;
  contactBeforeAction: boolean;
  problem_Id: number;
  change_Id: number;
  watchDate: number;
  finishingDate: string;
  finishingDescription: string;
  sms: boolean;
  updateNotifierInformation: boolean;
  addFollowersBtn: boolean;
  planDate: string;
  causingPartId: number;
  registrationSource: number;
  isActive: boolean;
  costCentre: string;
  isAbout_ReportedBy: string;
  isAbout_PersonsName: string;
  isAbout_PersonsEmail: string;
  isAbout_PersonsPhone: string;
  isAbout_PersonsCellPhone: string;
  isAbout_UserSearchCategory_Id: number;
  isAbout_Region_Id: number;
  isAbout_Department_Id: number;
  isAbout_OU_Id: number;
  isAbout_CostCentre: string;
  isAbout_Place: string;
  isAbout_UserCode: string;
  overWritePopUp: boolean;
  finalAction: number;
}